code refactored after review
changed namespace directions and made some improvements
